requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
Pillow==9.3.0
pdf2image==1.16.3
